Fix the BinaryIO problem.
There seemed to be a problem with multiple methods having the same
type erasure. Perhaps this has been preventing compilation? I don't
know. But fixing it seems to have kicked the variable threshold back
a functioning state. I should experiment with the other types of threshold.
But, the first thing I need to do is run this new setup on good/bad examples
and see how the score calculation is working.