Document EncodeData class.
There seems to be some unnecessary aspects to this code. I think it
would be better to pass the encoder in as a constructor parameter to
allow more control over the system. This would allow the encoder to be
specified both from file (using Encoder.getInstance) and through the
normal constructor style.